create blueprint for views and auth
## views.py

We store the standard routes for our website, where users can actually go to (e.g. homepage...). the login page will go into auth.py but any other page that the user can access will be put into views.py. This page is a blueprint which meaning it has a bunch of roots inside of it, it has a bunch of urls defined.

## auth.py

We also create a blue print here named auth which will serve as the blueprint for the authentication process.